projects/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Samuel Hinton", email = "samuelreay@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-sdk>=1.30.0",
    "prometheus-client>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]